backend/requirements.txt
fastapi
uvicorn
kafka-python
neo4j
python-dotenv
asyncio
